first draft of flux-burst-compute-engine
I need to build the images, and test an external burst (logic to be added).
If we are doing an external burst, we will need a way to submit jobs to
the external cluster, and maybe we can install the flux restful API with
a generated credential (will need a custom startup script). We also need
to add logic that handles saving the clusters by name/size. This is currently
in development mode, so I am mocking a job to test the basic interactions
with terraform

Signed-off-by: vsoch <vsoch@users.noreply.github.com>